rename lstm.py and loss_calc.py
