+++ test/test_beam_search.py
@@ -9,8 +9,8 @@ from xnmt.bridges import CopyBridge
 from xnmt.decoders import AutoRegressiveDecoder
 from xnmt.embedders import SimpleWordEmbedder
 from xnmt.input_readers import PlainTextReader
-from xnmt.lstm import UniLSTMSeqTransducer, BiLSTMSeqTransducer
-from xnmt.loss_calc import AutoRegressiveMLELoss
+from xnmt.recurrent import UniLSTMSeqTransducer, BiLSTMSeqTransducer
+from xnmt.loss_calculators import AutoRegressiveMLELoss
 from xnmt.param_collections import ParamManager
 from xnmt.transforms import NonLinear
 from xnmt.scorers import Softmax
