+++ xnmt/classifiers.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from xnmt import batchers, embedders, infererences, input_readers, losses, lstm, model_base, output, scorers, transducers, transforms
+from xnmt import batchers, embedders, infererences, input_readers, losses, recurrent, model_base, output, scorers, transducers, transforms
 from xnmt.persistence import serializable_init, Serializable, bare
 
 class SequenceClassifier(model_base.ConditionedModel, model_base.GeneratorModel, Serializable, model_base.EventTrigger):
@@ -26,7 +26,7 @@ class SequenceClassifier(model_base.ConditionedModel, model_base.GeneratorModel,
                src_reader: input_readers.InputReader,
                trg_reader: input_readers.InputReader,
                src_embedder: embedders.Embedder = bare(embedders.SimpleWordEmbedder),
-               encoder: transducers.SeqTransducer = bare(lstm.BiLSTMSeqTransducer),
+               encoder: transducers.SeqTransducer = bare(recurrent.BiLSTMSeqTransducer),
                inference=bare(infererences.IndependentOutputInference),
                transform: transforms.Transform = bare(transforms.NonLinear),
                scorer: scorers.Scorer = bare(scorers.Softmax)):
