+++ xnmt/translators.py
@@ -14,8 +14,8 @@ from xnmt.events import register_xnmt_handler
 from xnmt import infererences, input_readers, model_base
 from xnmt.input import Input, SimpleSentenceInput
 from xnmt.losses import FactoredLossExpr
-from xnmt.loss_calc import LossCalculator
-from xnmt.lstm import BiLSTMSeqTransducer
+from xnmt.loss_calculators import LossCalculator
+from xnmt.recurrent import BiLSTMSeqTransducer
 from xnmt.output import TextOutput, Output, NbestOutput
 from xnmt.persistence import serializable_init, Serializable, bare
 from xnmt.search_strategies import BeamSearch, SearchStrategy
